Have some images display. Update repo name.
Update image references and add gatsby-remark-image to have some of the
images display. Most of them don't at the moment, as many of the
references were reliant on Jekyll to display.

Moving my website over is less of a sure thing and more of an experiment
now. I don't have the CSS experience necessary to do a lot of the things
that I want to do with Gatsby. Update the repo name to reflect this.
I'll probably slowly work on this and see what I can do.